Rename aryl halide benchmarks
